webapi/tct-webmessaging-w3c-tests/webmessaging-py/w3c/Transferred_objects_unusable.htm

<!DOCTYPE html>
<html>
<head>
<title> Transferred objects are no longer usable on the sending side </title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/server.js?pipe=sub"></script>
</head>
<body>
<div id=log></div>

<script>
    setup({timeout:500});

    var description = "Test Description: " +
                      "Objects listed in transfer are transferred, not just cloned, meaning that they are " +
                      "no longer usable on the sending side.";

    var t = async_test(description);

    var DATA = {test: "e.ports[0].postMessage('TRANSFERRED')"};
    var ExpectedResult = ["PING", "TRANSFERRED"];
    var ActualResult = [];

    var TARGET;

    var flag = false;

    function PostMessageTest()
    {
        TARGET = document.querySelector("iframe");
        if (flag) {
            test(function()
            {
                assert_own_property(window, "MessageChannel", "window");

                var channel = new MessageChannel();

                channel.port2.onmessage = t.step_func(VerifyResult);

                channel.port1.postMessage("PING");

                TARGET.contentWindow.postMessage(DATA, "*", [channel.port1]);

                channel.port1.postMessage("PONG");

            }, "MessageChannel is supported.");
        }
      flag = true;
    }

    function VerifyResult(e)
    {
        ActualResult.push(e.data)

        if (ActualResult.length >= ExpectedResult.length)
        {
            assert_array_equals(ActualResult, ExpectedResult, "ActualResult");
            t.done();
        }
    }
</script>

<div style="display:none">
    <iframe width="70%" onload="PostMessageTest()" src=""></iframe>
</div>

<script>
        path = "http://" + __SERVER__NAME + ":" + __CORS__PORT + "/tests/webmessaging/w3c/support/ChildWindowPostMessage.htm";
        TARGET.src = path;
</script>
</body>
</html>
